Renames /operations to /generic
