+++ packages/ae-core/generic.js
@@ -1,0 +1,2 @@
+
+module.exports = require('./build/generic');
