--- packages/ae-core/operations.js
@@ -1,2 +1,0 @@
-
-module.exports = require('./build/operations');
