+++ packages/ae-docs/pages/homepage.mdx
@@ -18,6 +18,6 @@ yarn add @algebraic-effects/core @algebraic-effects/effects
 ### Import and use it
 ```js
 import { createEffect } from '@algebraic-effects/core'; /*core*/
-import { sleep, race } from '@algebraic-effects/core/operations'; /*generic effects*/
+import { sleep, race } from '@algebraic-effects/core/generic'; /*generic effects*/
 import { Exception, State } from '@algebraic-effects/effects'; /*State effect and Exception effect*/
 ```
