+++ packages/ae-effects/test/State.test.js
@@ -1,6 +1,6 @@
 
 import State from '../src/State';
-import { sleep, call } from '@algebraic-effects/core/operations';
+import { sleep, call } from '@algebraic-effects/core/generic';
 
 describe('State', () => {
   const logFn = jest.fn();
